LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-fr.html

<!DOCTYPE html><!-- webkit-test-runner [ language=fr ] -->
<html>
<head>
<script src="../../../../../resources/js-test.js"></script>
</head>
<body>
<script>
shouldBeEqualToString("navigator.language", "fr-fr");
</script>
</body>
</html>
