LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-fr-expected.txt

PASS navigator.language is "fr-FR"
PASS successfullyParsed is true

TEST COMPLETE
